CB-4054 no strictly related, but forced termination was missing some cleanup, moved everything to finalizeTermination. (Will conflict with master)
